requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
openpyxl==3.1.5
WTForms==3.2.2